+++ src/insertRule.js
@@ -5,6 +5,7 @@ require('./array-map');
 
 var camel = /([a-z])([A-Z])/g;
 var hyphens = '$1-$2';
+var contexts = {};
 
 function parseStyles (styles) {
   if (typeof styles === 'string') {
@@ -19,13 +20,62 @@ function parseStyles (styles) {
   }).join(';');
 }
 
-module.exports = function (selector, styles) {
-  var css = parseStyles(styles);
-  var sheet = document.styleSheets[document.styleSheets.length - 1];
-  var key = sheet.cssRules ? sheet.cssRules: sheet.rules;
-  if (sheet.insertRule) {
-    sheet.insertRule(selector + '{' + css + '}', key.length);
-  } else if (sheet.addRule) {
-    sheet.addRule(selector, css, key.length);
+function context (name) {
+  if (contexts[name]) {
+    return contexts[name];
   }
-};
+  var cache;
+  var rules;
+  var remove;
+
+  function getStylesheet () {
+    if (cache) {
+      return cache;
+    }
+    var style = document.createElement('style');
+    document.body.appendChild(style);
+    style.setAttribute('data-context', name);
+    cache = document.styleSheets[document.styleSheets.length - 1];
+    rules = cache.cssRules ? 'cssRules' : 'rules';
+    remove = cache.removeRule ? 'removeRule' : 'deleteRule';
+    return cache;
+  }
+
+  function add (selector, styles) {
+    var css = parseStyles(styles);
+    var sheet = getStylesheet();
+    var len = sheet[rules].length;
+    if (sheet.insertRule) {
+      sheet.insertRule(selector + '{' + css + '}', len);
+    } else if (sheet.addRule) {
+      sheet.addRule(selector, css, len);
+    }
+  }
+
+  function remove (selector) {
+    var sheet = getStylesheet();
+    var length = sheet[rules].length;
+    var i;
+    for (i = length - 1; i >= 0; i--) {
+      if (sheet[rules][i].selectorText === selector) {
+        sheet[remove](i);
+      }
+    }
+  }
+
+  function clear () {
+    var sheet = getStylesheet();
+    while (sheet[rules].length) {
+      sheet[remove](0);
+    }
+  }
+
+  add.clear = clear;
+  add.remove = remove;
+  contexts[name] = add;
+  return contexts[name];
+}
+
+var ctx = context('default');
+ctx.context = context;
+module.exports = ctx;